nm.profile: rewrite NM profile activation
This patch is rewriting the NM profile activation in order to use UUID
to identify which profile should be activated instead of using connection
id or device.

In addition, the ConnectionProfile class is being renamed to Profile and moved
to profile.py. A following PR will rewrite nm.active_connection.py into
profile.py in order to unify the NM profile management.

Signed-off-by: Fernando Fernandez Mancera <ffmancera@riseup.net>