tests: fix error on test_ovs_remove_port
In the ovs_remove_port test we are using nmcli directly using cmdlib.
nmcli need a second to refresh the information if not, it will be
outdated and then the test will not raise the exception.

This was not happening before because Nmstate was waiting 5 seconds
after applying the state.

Signed-off-by: Fernando Fernandez Mancera <ffmancera@riseup.net>