travis: add job testing tier1 tests with openvswitch2.13
This patch is creating a Travis CI job running tier1 tests using
openvswitch2.13. In order to do this, the container needs to uninstall
ovs 2.11 and then install ovs 2.13 if not there could be conflicts.

In addition, as Travis CI is expanding the environment variables, we
have created a new global variable and we need to use it directly in the
script call. This allowing us to introduce a long argument with a spaced
string, avoiding random scaping/expanding.

Signed-off-by: Fernando Fernandez Mancera <ffmancera@riseup.net>