nm ovs: Support OVS port using the same name of OVS interface
Previously, we are using interface name to determined the ports of OVS
bridge. This become a problem when the OVS port profile is sharing the
name with OVS interface.

This patch changed it to use `get_slaves()` of `NM.DeviceOvsBridge` and
`NM.DeviceOvsPort` to clarify this indistinct.

Also removed unused functions.

Signed-off-by: Gris Ge <fge@redhat.com>