ovsdb plugin: Don't add new interface
When OVS bridge is unmanaged by NetworkManager, NetworkManager plugin
will not include any information regarding it, the nispor plugin
neither(ovs bridge is user space only interface). The ovsdb plugin will
show it name and type.
This will cause oVirt crash when try to access OVS bridge information of
it which nmstate-0.3 does not.

To fix it, introduced `NmstatePlugin.is_supplemental_only` which
indicate whether plugin can only amend existing interface or not.

Set the ovsdb as supplemental only so it will never add new interface to
the `libnmstate.show()`.

Signed-off-by: Gris Ge <fge@redhat.com>