ovs: Ignore OVS capabilities at get interfaces
At containerize nmstate we cannot run "systemctl openvswitch status" so
even with a openvswitch running at host it will appear as non running
but the ovs information will arrive from NM dbus interface. This breaks
nmstatectl show since ovs-port is not part of nmstate and is not
included in the schema. This change remove the OVS check so ovs-port
is always excluded.

Signed-off-by: Quique Llorente <ellorent@redhat.com>