ifaces: do not remove unmanaged orphan interfaces
If there are unmanaged OVS interface present in the network state, NM
may report uncomplete information. Therefore, nmstate could consider
them as orphan and remove it when modifying the network state.

In order to fix this, nmstate should not consider it orphan and remove
it when it is not on desired state or marked as changed.

Signed-off-by: Fernando Fernandez Mancera <ffmancera@riseup.net>