sriov: Use verification retry to wait VF been created
When reactivating i40e interface with SR-IOV enabled, the kernel
takes some time(1 seconds or more in my test) to get the VF interface
ready in kernel. So at the time of libnmstate returns with success, the
VF interface might not be ready for use yet.

To fix that, we include VF interfaces in desire state when PV is
changed/desired. The verification retry will wait the VF to be ready for
use.

Unit test case and integration test case included.

Also fixed SRIOV integration test cases which are now all passing on i40e
NIC.

To test on real SRIOV NIC:

    cd tests/integration/
    sudo env TEST_REAL_NIC=ens1f1 pytest-3 sriov_test.py -vvv

Signed-off-by: Gris Ge <fge@redhat.com>