veth: retry import device when activating new veth peer
NetworkManager takes some extra time on creating the device object for
the new veth peer. Trying to activate an interface without importing the
device will end up on NetworkManager trying to create a new device for
the interface and that will end up on an error.

In order to fix this, Nmstate will retry to import the device up to 5
times with an interval of 0.5 seconds. If the device is not imported
during this time, then nmstate will try to activate it anyway.

Signed-off-by: Fernando Fernandez Mancera <ffmancera@riseup.net>