ovs: do not set ovs-bridge as master if duplicate names are present
When OVS interfaces are present with duplicate names i.e ovs-bridge,
ovs-interface and ovs-port with the same name, Nmstate is merging the
information to show it.

As Nmstate needs to store the DNS configuration as part of the
configuration of an interface, When the ovs-bridge is being reported
first than the ovs-interface, the IP configuration is being removed
internally. If the DNS configuration was stored on the ovs-interface, it
will fail because the IP configuration is required.

This solution only solves the following error and does not provide
duplicate names support.

```
  File "/usr/bin/nmstatectl", line 11, in <module>
    load_entry_point('nmstate==0.3.4', 'console_scripts', 'nmstatectl')()
  File "/usr/lib/python3.6/site-packages/nmstatectl/nmstatectl.py", line 67, in main
    return args.func(args)
  File "/usr/lib/python3.6/site-packages/nmstatectl/nmstatectl.py", line 267, in apply
    args.save_to_disk,
  File "/usr/lib/python3.6/site-packages/nmstatectl/nmstatectl.py", line 289, in apply_state
    save_to_disk=save_to_disk,
  File "/usr/lib/python3.6/site-packages/libnmstate/netapplier.py", line 71, in apply
    net_state = NetState(desired_state, current_state, save_to_disk)
  File "/usr/lib/python3.6/site-packages/libnmstate/net_state.py", line 58, in __init__
    self._ifaces.gen_dns_metadata(self._dns, self._route)
  File "/usr/lib/python3.6/site-packages/libnmstate/ifaces/ifaces.py", line 408, in gen_dns_metadata
    iface_metadata = dns_state.gen_metadata(self, route_state)
  File "/usr/lib/python3.6/site-packages/libnmstate/dns.py", line 116, in gen_metadata
    "name servers: %s" % server
libnmstate.error.NmstateValueError: Failed to find suitable interface for saving DNS name servers: 8.8.8.8
```

Integration test added.

Ref: https://bugzilla.redhat.com/1939557

Signed-off-by: Fernando Fernandez Mancera <ffmancera@riseup.net>