net_state: mark ifaces as ignored earlier
Nmstate is not touching unmanaged interfaces if they are not being
included in the desired state. If an unmanaged interface has a default
gateway configured and we try to add a dns server, Nmstate will pick the
unmanaged interface because at that point it is not marked as ignored.

As the interface is being marked as ignored later, the changes are not
being applied and is failing on verification. In order to avoid this,
the unmanaged interfaces should be marked as ignored earlier so they are
not considered for DNS/Routes changes.

Ref: https://bugzilla.redhat.com/1944582

Signed-off-by: Fernando Fernandez Mancera <ffmancera@riseup.net>