nm: Sort activation in alphabet order explicitly
The activation order on bridge/bond ports determins their controler's
MAC address. The default NetworkManager is using alphabet order on
boot. So we should to do the same on profile activation.

Previously, the `libnmstate.show()` provides the sorted interface list.
And the `Ifaces.all_iface()` dictionary is using the inserting order.
So we have the activation sorted in alphabet order implicitly.

This patch is fixing nothing, just do the activation sorting explicitly.

Signed-off-by: Gris Ge <fge@redhat.com>