ethtool: support backwards compatibility with NM 1.26
NetworkManager 1.26 and 1.28 do not have full support of ethtool
options. If the user is using nmstate 1.1.0 and nispor 1.1.0 with
NetworkManager 1.26/1.28, they won't be able to edit the interface
because it will raise a NmstateValueError exception.

Instead, Nmstate will fail on verification if the property was modified
but not supported. This will allow the backwards compatibility.

Signed-off-by: Fernando Fernandez Mancera <ffmancera@riseup.net>