libnmstate: introduce show_secrets argument
Nmstate needs to handle properly the secrets from the interfaces when
reporting, logging, crashing and verifying. In order to do that, this
patch is introducing the support for `show_secrets` argument for
`libnmstate.show()` and `--show-secrets` option for `nmstatectl show`.

Nmstate will not report, log, show when failing on verification or show
when crashing the sensitive data by default. When the new option is
present, Nmstate will show them on these situations. This can be used
for debugging or applying the current configuration in another host.

For `apply` and `edit` functions, Nmstate will only report, log or show
sensitive data if the desired state contains sensitive data. If the
desired state does not contain sensitive data, Nmstate will not do it.

Nmstatectl manpage documentation has been added.

Integration test case added.

Signed-off-by: Fernando Fernandez Mancera <ffmancera@riseup.net>