nmstate-autoconf: introduce nmstate-autoconf support
nmstate-autoconf is a command line tool that uses libnmstate in order to
confgure the network state automatically using LLDP information. This
tool is experimental only.

By running `nmstate-autoconf`, it will identify the interfaces connected
to the same VLAN and will group them with a Bond interface, creating a
host VLAN in top of the bond.

The bond state will be the following:
```
- name: bond50
  type: bond
  state: up
  link-aggregation:
    mode: balance-rr
    port:
    - port0
    - port1
```

The VLAN state will be the following:
```
- name: prod-net
  type: vlan
  state: up
  vlan:
    base-iface: bond50
    id: 50
```

Integration test cases added.

Signed-off-by: Fernando Fernandez Mancera <ffmancera@riseup.net>