nispor: fix show of empty next_hop_address and destination
The correct way of representing an empty next_hop_address is using
"0.0.0.0" for IPv4 and "::" for IPv6. This configuration is valid
because an user should be able to specify the following routes:

```
---
routes:
  config:
  - destination: 0.0.0.0/0
    next-hop-address: 0.0.0.0
    next-hop-interface: dummy
  - destination: ::/0
    next-hop-address: "::"
    next-hop-interface: dummy

```

That means each of the hosts within the range of the route are
considered to be directly connected through that interface.

For example, using iproute2 the user should introduce the following
command:

`ip route 0.0.0.0 0.0.0.0 dummy`

Integration test case added.

Ref: https://bugzilla.redhat.com/1985879

Signed-off-by: Fernando Fernandez Mancera <ffmancera@riseup.net>