ovsdb: drop NM.connection.uuid entry
In first place, the user should not be able to modify the NM external_id
from the ovsdb. If the user save a current state with an ovs bridge and
then the NM.connection.uuid is modified by NM, the user won't be able to
apply the state saved.

This is fixing the following verification error:

```
difference
==========
--- desired
+++ current
@@ -51,4 +51,4 @@
 mtu: 1500
 ovs-db:
   external_ids:
-    NM.connection.uuid: 930aa309-99b1-4fe9-a300-ceec3efd9907
+    NM.connection.uuid: 2fe14d2b-3465-43bb-8a49-f65ded332238
```

Ref: https://bugzilla.redhat.com/1660250

Signed-off-by: Fernando Fernandez Mancera <ffmancera@riseup.net>