nm.connection: add postfixes to OVS bridges and interfaces connections
When configuring an OVS bridge and interface with the same using
Nmstate, it will generate two connection with the same name. If the user
is willing to use nmcli to manage this profiles, it is not user friendly
because they will need to use the connection UUID instead.

In order to solve this, Nmstate is adding a postfix on the connection
name. "-if" for OVS interface and "-br" for OVS bridge.

Please, note that this issue does not affect kernel interfaces because
it is not possible to have duplicated interfaces names in kernel.

Integration test case added.

Ref: https://bugzilla.redhat.com/1998218

Signed-off-by: Fernando Fernandez Mancera <ffmancera@riseup.net>