rust: add golang bindings
This patch is introducing golang bindings for the C bindings of
nmstate. The following functions are being introduced:

 - RetrieveNetState()
   * Returns (state, error) where state is a string containing the
     current state.

 - ApplyNetState(state string)
   * Returns (state, error) where state is a string containing the
     desired state applied.

 - CommitCheckpoint(checkpoint string)
   * Returns (checkpoint, error) where checkpoint is the string
     containing the checkpoint path committed.

 - RollbackCheckpoint(checkpoint string)
   * Returns (checkpoint, error) where checkpoint is the string
     containing the checkpoint path used for rolling back.

The logs are written to the io.Writer passed by the user. This way the
user can use the logs as they prefer.

Signed-off-by: Fernando Fernandez Mancera <ffmancera@riseup.net>
(cherry picked from commit 20085a5975d992198ee38cd3a6a9c408dc580909)