rust, ip: Fix deserialization/serialization of ipv4/ipv6
Nmstate is using a custom serializer for ipv4 and ipv6. We are relying
in `prop_list` to serialize the values but when creating a NetState
struct and doing `serde::to_string(net_state)` the `prop_list` is not
populated. Therefore it will not show the properties.

For example if we generate a NetState from the following file:

```yaml
---
interfaces:
  - name: eth1
    type: ethernet
    state: up
    ipv4:
      enabled: true
      address:
        - ip: 192.168.100.200
          prefix-length: 24
```

And we use the `serde::to_string()` method, it will show:

```yaml
---
interfaces:
  - name: eth1
    type: ethernet
    state: up
    ipv4:
      enabled: true
```

In order to fix this, we must implement our deserializer in which we
populate `prop_list`. This must be done on each struct that relies on
`prop_list` to serialize.

Signed-off-by: Fernando Fernandez Mancera <ffmancera@riseup.net>