rust, ip: Fix deserialization/serialization of ipv4/ipv6
Nmstate is using a custom serializer for ipv4 and ipv6. We are relying
in `prop_list` to serialize the values but when creating a NetState
struct and doing `serde::to_string(net_state)` the `prop_list` is not
populated. Therefore it will not show the properties.

For example if we generate a NetState from the following file:

```yaml