rust nm: Ignore failure when deleting unmanaged interface
For preexist unmanaged software interface, the deletion of its profile
will not delete the interface in NM.
We have to delete it manually.

There is race problem when profile is deleting, the device might still
in managed state or gone at the time we invoke device deletion method.
So we just ignore the failure of device deletion and let verification
stage found the problem.

Signed-off-by: Gris Ge <fge@redhat.com>