rust: Set vlan-filtering 0 on port.vlan: {}
The vlan-filtering at linux-bridge is set to 0 if there is no vlan
configured at ports, but previously with the python version it was also
true if the vlan diciontary was empty. This change implement a fix for
that regression comparing with "default" struct.

Signed-off-by: Quique Llorente <ellorent@redhat.com>
(cherry picked from commit 2708db7f9fe5a74d903230bd986539beaadda9ea)