checkpoint: Extend rollback with timeout
The checkpoint rollback timeout has to be extended so it start to count
after apply is executed and some ovs operations are done, the extended
NM dbus operation is adding that "now" not to the checkpoint timeout.
This change set the extended time to timeout instead of choosing min
with will make checkpoint timeout be always the default 60s.

Signed-off-by: Enrique Llorente <ellorent@redhat.com>