nm: Generate stable UUID for keyfile
When user generate configuration for the same state, it is normal for
him to expect the same output every time. Hence we generate the same
UUID in that case.

To achieve that, we use UUID namespace URL using
"{iface_type}://{iface_name}" format to generate the stable UUID.
This might cause overlap UUID, so we only do this when there is no
existing NetworkManager profiles/connections.

Integration test case included.

Signed-off-by: Gris Ge <fge@redhat.com>