nm: Retry on verification failure
When create a VRF containing both new bond and new bond VLAN, nmstate
has 50% chance of verification failure. Example yaml:

```yaml
---
interfaces:
- name: bond0.100
  type: vlan
  vlan:
    base-iface: bond0
    id: 100
- name: bond0
  type: bond
  link-aggregation:
    mode: balance-rr
- name: vrf0
  type: vrf
  state: up
  vrf:
    port:
    - bond0
    - bond0.100
    route-table-id: 1000
```

The root cause of this is NM somehow forgot to activate the bond VLAN as
instructed by `connection.autoconnect-slaves`, this lead to verification
failure due to missing VRF ports.

Nmstate is trusting NM on this `autoconnect` on ports, so nmstate
skipped the manual activation on bond VLAN as its controller vrf0 is
has manually activated. Apparently, NM has race problem when this bond
VLAN is depending on another new bond.

We can fix this issue by manually activating the port connection if it has
parent also. But we may get more race problem like this in the future.
So we do one time retry on apply() if verification failed.

Integration test case included.

Signed-off-by: Gris Ge <fge@redhat.com>