nm ovs: Fix race problem when attaching bond to OVS bridge
Got failure when attaching linux bond to ovs bridge:

    libnmstate.error.NmstateLibnmError: Activate profile
    uuid:cf81a923-5999-4a21-bd7e-dc96c7977451 iface:eth1 type: ethernet
    failed: reason=<enum
    NM_ACTIVE_CONNECTION_STATE_REASON_DEVICE_DISCONNECTED of type
    NM.ActiveConnectionStateReason><enum
    NM_DEVICE_STATE_REASON_DEPENDENCY_FAILED of type NM.DeviceStateReason>

This is caused by race between we activating OVS port profile manually
and NettworkManager `autoconnect-slaves` function.

To fix the problem, for OVS bridge and OVS port, the `IP_CONFIG` state
is not considered as activated as they are not allowed to have IP.

Integration test case included.

Signed-off-by: Gris Ge <fge@redhat.com>