bond: Support integer option
Supporting integer for these bond options:
 * `mode`
 * `all_slaves_active`
 * `arp_all_targets`
 * `arp_validate`
 * `fail_over_mac`
 * `primary_reselect`
 * `xmit_hash_policy`
 * `ad_select`
 * `lacp_rate`

The reason our unit test case failed to catch this problem is because
`serde_yaml` automatically convert integer to string.
But in `Interface`, it was deserialize by `serde_json` which
implantation does not have auto convert on int to string.

Unit test case expanded.

Signed-off-by: Gris Ge <fge@redhat.com>