nm: Fix moving bridge port to bond
When moving bridge port with VLAN filtering setting to bond, we got
failure:
    A connection with a 'bridge-port' setting must have the slave-type
    set to 'bridge'. Instead it is 'bond'

The root cause is we forgot to remove `NmSettingBridgePort` when
detaching port away from bridge.

The fix is remove `NmSettingBridgePort` and
`NmSettingOvsBridgeInterface` if not bridge port or OVS interface
anymore.

Integration test cases for linux bridge and OVS bridge are included.

Signed-off-by: Gris Ge <fge@redhat.com>