nm sriov: Do not touch SR-IOV if not desired
We should not create SRIOV settings in NetworkManager if that is not
desired.

In the verification stage, we do not verify SR-IOV VFS if not desired.

Unit test case added to confirm verification will not check SRIOV if
not desired.

Integration test case included.

Manual test been done on Intel X710.

Signed-off-by: Gris Ge <fge@redhat.com>