nm: Support `dns-data` in NetworkManager 1.41
The NetworkManager 1.41 introduced `dns-data` and deprecating `dns` DBUS
propriety.

Without this patch, we will not able to purge DNS on NM 1.41
as we always merge unknown DNS propriety which in this case `dns-data`.
The `dns-data` is effective if `dns` is not mentioned, hence we cannot
purge DNS in NM 1.41+.

We cannot simply set empty array in `dns` which take priority over
`dns-data` when applying because IPv4 and IPv6 dns setting has different
dbus signature which require at lease one dns server to determine.

This patch only enable the querying support on `dns-data`, but still use
old `dns` property when applying to achieve backwards compatibility,
and drop `dns-data` when applying also.

The integration test case `test_dns_edit_nameserver_with_static_gateway`
reproduced the problem and could confirm the fix.

Signed-off-by: Gris Ge <fge@redhat.com>