route_rule: support `from all to all` routing policy
Add support to `from all to all` routing policy. In static config
generator it will add `from 0.0.0.0/0` or `from ::/0` to the keyfile.

A new parameter `family` have been introduced to the route-rule config
field. This parameter is used to specify the address family. If `family`
is not specified it will fail on validation, if not specified from
Nispor it will assume IPv4 as it was done in the past.

Example:

```yaml
route-rules:
  config:
    - route-table: 254
      priority: 100
      family: ipv4
```

Integration test cases added.

Fixes #1417

Signed-off-by: Fernando Fernandez Mancera <ffmancera@riseup.net>