Makefile: add `-Z namespaced-features` to fix compiler error
When compiling nmstatectl in Copr infraestructure, it is failing with an
error related to the `Cargo.toml`. This option has been already
stabilized but Copr is using an older version of rustc.

```
namespaced features with the `dep:` prefix are only allowed on the
nightly channel and requires the `-Z namespaced-features` flag on the
command-line
```

Signed-off-by: Fernando Fernandez Mancera <ffmancera@riseup.net>