nm: Improve error retry
In general, we have three types of retry:
 1. NM activation retry
 2. Nmstate verification retry with pure wait.
 3. Nmstate verification retry by apply the same state again.

Only the 1) NM activation retry in changed by this patch.
With this patch we will retry only the __failed__ connection with
retryable error after increasing(1,2,4,8,16,32 seconds) interval.

Removed the retry at `NmApi` level base on test results.

To identify which `NmError` is retryable, this patch expanded the
`NmError` to cover detailed error class:
 * `NmManagerError`
 * `NmSettingError`
 * `NmConnectionError`

The `NmError` will convert to `NmstateError` for future processing.

Only these `NmstateError` are retryable:

 * `ErrorKind::PluginFailure`
 * `ErrorKind::Bug`
 * `ErrorKind::VerificationError`

Tested for 500 VLAN + 500 bridge, finished in 5 minutes.
Tested with OVS plugin uninstalled.
Tested with loopback not supported.

Signed-off-by: Gris Ge <fge@redhat.com>