state: Introduce crate level API - `MergedNetworkState`
Rewrite the state merging process of nmstate into `MergedNetworkState` which
contains:
 * `MergedHostNameState`
 * `MergedDnsState`
 * `MergedInterfaces`
 * `MergedOvsDbGlobalConfig`
 * `MergedRoutes`
 * `MergedRouteRules`

The most important `MergedInterfaces` performs actions in the order of:
 1. `MergedInterface.pre_inter_ifaces_process()` for actions required to
    be done before follow up. For example, empty OVS bridge should have
    a OVS internal interface auto created.
 2. `MergedInterfaces.process()` for actions require knowledge of
    multiple interfaces.
 3. `MergedInterface.post_inter_ifaces_process()` for actions require
    single interface desire and current status.
 4. `Interface.sanitize()` for actions apply to `for_apply` and
    `for_verify`.
 5. `Interface.sanitize_for_verify()` for action only for verify.

The `MergedInterface` has these three data stored:
 * `merged`: Contains both desired and current data.
 * `for_apply`: Optionally for state to apply.
 * `for_apply`: Optionally for state to verify.

With this design, when processing desire state, we can have clear
knowledge on value source and be explicit on the type of
modification(for apply or for verify).

Also moved the ifaces-based route/route_rule/dns code to `nm` folder as
they are NetworkManager specific as:
 * `nm::dns::store_dns_config()`
 * `nm::route::store_route_config()`
 * `nm::route_rule::store_route_rule_config()`

In the future, we might expose `MergedNetworkState` out for `dry-run` or
pre-edit validation.

All unit tests and integration tests passed.

Signed-off-by: Gris Ge <fge@redhat.com>