nm: Fix error when bring down unmanaged interface
When set `state: down` for unmanaged interface, nmstate will fail on
verification. This is because nmstate did not convert the interface to
managed before deactivate the profile.

The fix is create profile for unmanaged down interface, activate it then
deactivate it, for virtual interface, delete the link also.

Integration test cases included.

Signed-off-by: Gris Ge <fge@redhat.com>