nm: Retry when NetworkManager starting up
When OS boot up with `nmstate.service` enabled, nmstate will fail as
NetworkManager daemon is not fully setup the DBUS interface yet.

The `nmstate.service` has `After=NetworkManager.service`. Considering
`NetworkManager.service` is dbus type systemd service, it is guaranteed
that when `nmstate.service` starts, the DBUS interface is acquired by
NetworkManager. The failure is
`org.freedesktop.DBus.Error.UnknownMethod` as NetworkManager has not
finished the DBUS interface setup yet.

The `org.freedesktop.DBus.Error.UnknownMethod` is too generic, we cannot
assume it as daemon not started with checking the error string.
Considering the DBUS error string is not an API, we will not parse that
error message. Instead, we change top level retry to cover on querying
failures in `apply()`. The retry has been expend to twice with 2 seconds
interval(was 1 second interval).

The `nmstatectl service` command will sleep 2 seconds to avoid the
meaningless retry.

Do not want to create complex integration test case for this corner case.
Manually tested.

Signed-off-by: Gris Ge <fge@redhat.com>