ip: Preserver IP when converting unmanaged interface to managed
When converting unmanaged(or external managed) interface to managed,
nmstate will sort the IP address in alphabet order which is incorrect
and break user's environment.

The root cause of this is nispor plugin sort the IP address during querying,
which lead to the IP address sending to NM been sorted.

The fix is just remove the IP address sorting  in nispor plugin.

Integration test case been included.

Signed-off-by: Gris Ge <fge@redhat.com>