nmstate.service: remain after exit
This is generally a good idea for oneshot units.  If we don't remain after
exit, the unit will run and then immediately stop itself.  As a result, if
a dependent unit is started in a subsequent transaction, nmstate.service
will run again.  Furthermore, if the service is stopped, systemd will
consider any dependent targets to no longer be running.

Signed-off-by: Benjamin Gilbert <bgilbert@redhat.com>