dns: Support static DNS search with auto DNS nameserver
Introduce the support of static DNS search along with DNS nameserver
learn from DHCP/autoconf.

For implementation in NM, we try this order to find a interface to store
the DNS config:
 * Desired interface which currently hold DNS config and still valid for
   DNS.
 * Auto interface, prefer desired interface.
 * IP enabled interface, prefer desired interface.

We will not use global DNS for this use case
as it will suppress DNS nameserver learn from DHCP.

Integration test case included.

Signed-off-by: Gris Ge <fge@redhat.com>