cli: Simplify and make consistent logging in `persist-nic-names`
In the pinning path, we used "would", which would be more appropriate
for dry run mode, but we also used it in non-dry run mode. In the
cleanup path, we used "will" only in dry run mode.

Since "will" works for both dry run mode and non-dry run mode, tweak
things so we always say "will". This simplifies things since we don't
need different logging for each mode.

In the cleanup path, log when we actually remove the link file for
consistency with the pinning path.

While we're here, merge the kernel arg and link file logging.

Signed-off-by: Jonathan Lebon <jonathan@jlebon.com>
Signed-off-by: Colin Walters <walters@verbum.org>