cli: Readd `perist-nic-names --inspect`, rework internals
Today the logs from this invocation may end up either
in the systemd journal or in a container log.

And after that one run, they will quickly get lost back in the
shuffle.

The original motivation behind `--inspect` is to be able
to quickly and conveniently see the state of the system; it's
something we can run every time the OpenShift machine-config-daemon
starts up on a node.

That way, we'll always be able to look at the logs and see what
happened with NIC persistence.

In the implementation of this, rework things such that "dry_run"
is a bool passed down instead of augmenting the action.

This makes the code flow much more clearly; instead of having
an early return in `run_persist_immediately` if the action is
to cleanup we separate the persist action into its own function
too, and then just dispatch to one or the other.

Signed-off-by: Colin Walters <walters@verbum.org>