API: add ovn subtree
This commit adds a new subtree - `ovn` - to the `NetworkState`. This
feature requires the ovsdb module enabled, since it is essentially just
syntactic sugar on top of the ovsdb `ovn-bridge-mappings` external ID
key.

To add a new bridge-mapping the user should apply the following snippet:
```yaml
ovn:
  bridge-mappings:
  - localnet: tenantblue
    bridge: ovsbr1
    state: present
```

This API supports partial editing; thus, to delete a particular mapping,
the user **must** set its state to absent:
```yaml
ovn:
  bridge-mappings:
  - localnet: tenantred
    bridge: ovsbr1
    state: absent
```

Keep in mind the `bridge` attribute is optional when deleting.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>