ovn: do not allow for duplicate localnet keys
This commit makes it impossible for the user to provide multiple
`OvnBridgeMappings` with the same `localnet` key in the desired state.

The way for updating a key is to simply refer to it once, and provide
the update `bridge` value. When doing that, the `state` attribute is not
required. Refer to the following snippet for an example (assuming an
existing `net1: br1` mapping):

```yaml
ovn:
  bridge-mappings:
  - localnet: net1
    bridge: new-bridge-123
```

The example above will change the `net1` `localnet` mapping to the
`new-bridge-123` OVS bridge.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>