nm: do not attach ovs-bridge to itself when creating a profile
If ovs-bridge and ovs-interface shares name and ovs-interface is
modified, during the creation of the ovs-bridge profile we are setting
itself as a controller.

That is wrong and NetworkManager is reporting the following error:

```
libnmstate.error.NmstateLibnmError: Update profile
uuid:ba206f8f-2ed6-486d-a339-9d1f62c5cb84 iface:br1 type:ovs-bridge
failed with error=nm-connection-error-quark: connection.slave-type:
Cannot set 'master' without 'slave-type' (6)
```

In order to solve that, before setting the controller we check that it
is not itself.

Signed-off-by: Fernando Fernandez Mancera <ffmancera@riseup.net>