bond: Add support for the `ports-config`
The `ports-config` is the field to specify a list of ports
configuration, the supported settings are `priority` and `queue-id`.
The bond port priority is only supported to set when bonding mode is
active-backup(1) or balance-tlb (5) or balance-alb (6).

Please note that the `port` field is not required when specifying the
`ports-config`. An InvalidArgument error will be raised during apply
when user also specifies the `port` and the port names specified in
`port` conflict with the port names specified in `ports-config`.

Example:

```
interfaces:
  - name: bond0
    type: bond
    state: up
    link-aggregation:
      mode: active-backup
      options:
        miimon: 140
      ports-config:
        - name: eth1
          priority: -1
          queue-id: 0
        - name: eth2
          priority: 2
          queue-id: 1
```

Signed-off-by: Wen Liang <liangwen12year@gmail.com>