ip: Support treating string as int for `prefix-length`
When the network role user is using the `network_state` variable to
configure the network, and if they are using Jinja2 template to
define the `prefix-length`, the type conversion
`prefix-length: "{{ __str_val | int }}"` does not work as expected, the
type for `prefix-length` in the end is still string. Therefore, nmstate
need to support treating string as int for `prefix-length` in order to
make the apply succeed.

Signed-off-by: Wen Liang <liangwen12year@gmail.com>