ipsec: Support more options
Add support of these options:
 * `ikelifetime`
 * `salifetime`
 * `ike`
 * `esp`

Please refer to libreswan document for definition of them.

Integration test case included for `ikelifetime` and `salifetime`.

For `ike` and `esp`, hard coding them require effort on syncing it with
OS encryption method of OS and ipsec. So we are trusting user know how
to set these options correct, no integration test case for it.

Signed-off-by: Gris Ge <fge@redhat.com>