lib: Add enum variants missing serde annotations
At some enums the serde serialization is done using custom functions
instead of annotations, this change keep the display functions but
implement the serde serialization with the proper annotations.

This helps with code generators since traversing a well known annotation
is more appropiate than hardcode or traverse a function with random
code.

Signed-off-by: Enrique Llorente <ellorent@redhat.com>