connection: drop "controller" from the extra values
NetworkManager is introducing a new alias for "master" which is
"controller". Nmstate is taking unknown DBus parameters and sending it
back unmodified. NetworkManager is prioritising "master" over
"controller" if both are present but when removing the port only
"controller" is present in Nmstate code.

In order to fix this temporarily, drop "controller" from the other
values dictionary. This is going to be dropped when using "controller"
everywhere.

Signed-off-by: Fernando Fernandez Mancera <ffmancera@riseup.net>