lib, bond: Move numeric variants to serde alias
All the bond enums implement a TryFrom to convert NumberOrString and
support both the numeric and identity variants. This change move that to
the struct using serde "alias" feature and also use "remote" to be able
to use a custom serializer without loosing the originaed from serde.

This also helps with the fugure golang code generator since validation
annotations can be generated from serde annotations at the enum
variants.

Signed-off-by: Enrique Llorente <ellorent@redhat.com>